Resolves build issue
