+++ web/beacon-app/src/features/auth/routes/VerifyPage.tsx
@@ -37,7 +37,7 @@ function VerifyPage() {
                 <p className="mt-6">Thank you for verifying your email address.</p>
                 <p className="mt-6">Log in now to start using Ensign.</p>
               </div>
-              <Button color="secondary" size="large" className="mt-16" onClick={redirectToLogin}>
+              <Button variant="secondary" size="large" className="mt-16" onClick={redirectToLogin}>
                 Log in
               </Button>
             </div>
